move handleLine to its own file
